Fix stand_leg_right function
Changed the angle of the ankle motors to increase the odds of a
successful stand on right leg. Still highly dependent on the mechanical
structure though. Also changed the global variable increment to create
smoother motion for all the functions BEFORE issuing a walk. A bug
resulted in very jagged motions until the increment variable was
overwritten by a walk function.